functions/resize_photo/requirements.txt
google-cloud-storage==2.13.0
Pillow==10.1.0